requirements.txt
pytest==3.10.1
selenium==3.0.0
allure-pytest==2.3.1b1
allure-python-commons==2.3.1b1
requests==2.21.0
pytest-rerunfailures==7.0
webdriver-manager==1.7
pytest-fixture-config==1.4.0
logging==0.4.9.6
psycopg2==2.8.6